+++ src/core/maps/store/actionsMapsZoomTo.js
@@ -35,7 +35,7 @@ export default {
             }
             else if (id === "zoomToGeometry") {
                 urlValues = urlValues.split(",").map(value => value.toUpperCase().trim());
-                allowedValues = conf.allowedValues.map(value => String(value));
+                allowedValues = conf.allowedValues?.map(value => String(value));
                 // zoom to bezirk by urlParameter ?zoomtogeometry=1, means zoom to 1. entry in allowedValues
                 if (urlValues.length === 1 && allowedValues !== undefined && !allowedValues.includes(urlValues[0]) && parseInt(urlValues[0], 10) < allowedValues.length && parseInt(urlValues[0], 10) > 0) {
                     urlValues[0] = allowedValues[parseInt(urlValues[0], 10) - 1];
@@ -49,7 +49,7 @@ export default {
                 addFeatures = config.addFeatures;
             }
 
-            return featureProvider.getAndFilterFeatures(layerId, property, urlValues)
+            return featureProvider.getAndFilterFeatures(layerId, property, urlValues, {featureProjection: rootGetters["Maps/projection"]})
                 .then(featureCollection => {
                     let filteredFeatures = featureCollection;